sdk: tests: refactor: move pfs reducer unit tests
